Update README.md
I coded the entire Stamp Out Stigma web site, about a dozen pages; this code sample is for the splash page.  I was also the webmaster once the site went live.  I picked this piece because I want to highlight the neat coding and demonstrate the range of clients I am comfortable working with.